+++ CHANGELOG.md
@@ -75,6 +75,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ANC | Fix crashes due to invalid data
 - Engine | Remove the menu icon on the login screen
 - Engine: Fixes Structure map based resource extraction not working
+- eCBIS | Login page updated
 
 ### Changed
 
@@ -97,6 +98,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Refactor COVAX Immunization extraction to use StructureMap-based extractio
+- eCBIS | Side menu changed
 
 ## [0.0.8 EIR, 0.0.1 - ANC] - 2021-09-11